chore(suite): refactor bridge related troubleshooting tips to reduce lines of code
