+++ packages/suite/src/components/suite/troubleshooting/tips/index.tsx
@@ -3,8 +3,7 @@ import { isWeb, isLinux, isAndroid } from '@trezor/env-utils';
 import { Translation } from 'src/components/suite/Translation';
 import { isWebUsb } from 'src/utils/suite/transport';
 
-import { BridgeStatus } from './BridgeStatus';
-import { BridgeInstall } from './BridgeInstall';
+import { BridgeStatus, BridgeInstall } from './BridgeTip';
 import { UdevDescription } from './UdevDescription';
 
 export const TROUBLESHOOTING_TIP_BRIDGE_STATUS = {
